requirements.txt
annotated-types==0.6.0
backoff==2.2.1
certifi==2023.7.22
charset-normalizer==3.3.1
idna==3.4
psycopg2-binary==2.9.9
pydantic==2.4.2
pydantic-settings==2.0.3
pydantic_core==2.10.1
python-dotenv==1.0.0
requests==2.31.0
typing_extensions==4.8.0
urllib3==2.0.7
